refactor: moved table component for Transaction preview to a sub-component to mitigate render blockages while loading
