+++ client/src/pages/BudgetOverview/components/TransactionPreview/components/TPTable/TPTable.types.ts
@@ -1,0 +1,6 @@
+export interface IProps {
+	categoryId: number
+	endDate: string
+	setLoading: (loading: boolean) => void
+	startDate: string
+}
