+++ client/src/pages/BudgetOverview/components/TransactionPreview/components/TPTable/index.ts
@@ -1,0 +1,1 @@
+export { default } from './TPTable';
